add script for generating ids
